+++ dflib-csv/src/test/java/org/dflib/csv/CsvLoaderTest.java
@@ -1,9 +1,9 @@
 package org.dflib.csv;
 
+import org.apache.commons.csv.CSVFormat;
 import org.dflib.DataFrame;
 import org.dflib.ValueMapper;
 import org.dflib.junit5.DataFrameAsserts;
-import org.apache.commons.csv.CSVFormat;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
@@ -14,6 +14,7 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
+import static org.dflib.Exp.$col;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class CsvLoaderTest extends BaseCsvTest {
@@ -405,5 +406,79 @@ public class CsvLoaderTest extends BaseCsvTest {
                 .expectHeight(2)
                 .expectRow(0, null, "three")
                 .expectRow(1, "five", null);
+    }
+
+    @Test
+    public void noCompaction() {
+        DataFrame df = new CsvLoader()
+                .col("A", s -> s != null ? Integer.parseInt(s) : null)
+                .emptyStringIsNull()
+                .load(inPath("with_duplicates.csv"));
+
+        new DataFrameAsserts(df, "A", "B")
+                .expectHeight(6)
+                .expectRow(0, 1, "ab")
+                .expectRow(1, 40000, "ab")
+                .expectRow(2, 40000, "bc")
+                .expectRow(3, 30000, "bc")
+                .expectRow(4, 30000, null)
+                .expectRow(5, null, "bc");
+
+        DataFrame idCardinality = df.cols().select(
+                $col("A").mapVal(System::identityHashCode),
+                $col("B").mapVal(System::identityHashCode));
+
+        assertEquals(6, idCardinality.getColumn(0).unique().size());
+        assertEquals(6, idCardinality.getColumn(1).unique().size());
+    }
+
+    @Test
+    public void compactCol_Name() {
+        DataFrame df = new CsvLoader()
+                .compactCol("A", s -> s != null ? Integer.parseInt(s) : null)
+                .compactCol("B")
+                .emptyStringIsNull()
+                .load(inPath("with_duplicates.csv"));
+
+        new DataFrameAsserts(df, "A", "B")
+                .expectHeight(6)
+                .expectRow(0, 1, "ab")
+                .expectRow(1, 40000, "ab")
+                .expectRow(2, 40000, "bc")
+                .expectRow(3, 30000, "bc")
+                .expectRow(4, 30000, null)
+                .expectRow(5, null, "bc");
+
+        DataFrame idCardinality = df.cols().select(
+                $col("A").mapVal(System::identityHashCode),
+                $col("B").mapVal(System::identityHashCode));
+
+        assertEquals(4, idCardinality.getColumn(0).unique().size());
+        assertEquals(3, idCardinality.getColumn(1).unique().size());
+    }
+
+    @Test
+    public void compactCol_Pos() {
+        DataFrame df = new CsvLoader()
+                .compactCol(0, s -> s != null ? Integer.parseInt(s) : null)
+                .compactCol(1)
+                .emptyStringIsNull()
+                .load(inPath("with_duplicates.csv"));
+
+        new DataFrameAsserts(df, "A", "B")
+                .expectHeight(6)
+                .expectRow(0, 1, "ab")
+                .expectRow(1, 40000, "ab")
+                .expectRow(2, 40000, "bc")
+                .expectRow(3, 30000, "bc")
+                .expectRow(4, 30000, null)
+                .expectRow(5, null, "bc");
+
+        DataFrame idCardinality = df.cols().select(
+                $col("A").mapVal(System::identityHashCode),
+                $col("B").mapVal(System::identityHashCode));
+
+        assertEquals(4, idCardinality.getColumn(0).unique().size());
+        assertEquals(3, idCardinality.getColumn(1).unique().size());
     }
 }